Zend/zend_types: move `HashTable` and `HT_*` to `zend_hash.h`
