+++ Zend/zend_hash.c
@@ -18,6 +18,7 @@
    +----------------------------------------------------------------------+
 */
 
+#include "zend_hash.h"
 #include "zend.h"
 #include "zend_globals.h"
 #include "zend_variables.h"
